moved our module specific tests to own spec file
